move framework e2e test config into separate files
